fix: informer caching fails, if migration is in progress
