+++ pkg/service/HelmAppService.go
@@ -26,6 +26,7 @@ import (
 	registry2 "github.com/devtron-labs/common-lib/helmLib/registry"
 	k8sCommonBean "github.com/devtron-labs/common-lib/utils/k8s/commonBean"
 	k8sObjectUtils "github.com/devtron-labs/common-lib/utils/k8sObjectsUtil"
+	globalConfig "github.com/devtron-labs/kubelink/config"
 	"github.com/devtron-labs/kubelink/converter"
 	error2 "github.com/devtron-labs/kubelink/error"
 	"github.com/devtron-labs/kubelink/pkg/cache"
@@ -114,7 +115,7 @@ type HelmAppServiceImpl struct {
 	k8sService        K8sService
 	randSource        rand.Source
 	K8sInformer       k8sInformer.K8sInformer
-	helmReleaseConfig *HelmReleaseConfig
+	helmReleaseConfig *globalConfig.HelmReleaseConfig
 	k8sUtil           k8sUtils.K8sService
 	pubsubClient      *pubsub_lib.PubSubClientServiceImpl
 	clusterRepository repository.ClusterRepository
@@ -123,7 +124,7 @@ type HelmAppServiceImpl struct {
 }
 
 func NewHelmAppServiceImpl(logger *zap.SugaredLogger, k8sService K8sService,
-	k8sInformer k8sInformer.K8sInformer, helmReleaseConfig *HelmReleaseConfig,
+	k8sInformer k8sInformer.K8sInformer, helmReleaseConfig *globalConfig.HelmReleaseConfig,
 	k8sUtil k8sUtils.K8sService, converter converter.ClusterBeanConverter,
 	clusterRepository repository.ClusterRepository,
 	registrySettings registry2.SettingsFactory,
